SQLAlchemy 1.4 requires SQLALCHEMY_DATABASE_URI to use `postgresql` rather than `postgres` for the connection uri to the database.
When deploying to paas the database postgres environment variables are set using VCAP_SERVICES provided by PaaS. When we start up the app and set the properties we need to replace the postgres string with postgresql for the app to start up properly.
This wasn't caught locally or with the unit tests because we were setting this property with postgresql.