Republish gov.uk/alerts every night to clear down planned tests
We have made it so that gov.uk/alerts shows a ‘1 planned test’ banner
for the whole of the day when there has been an operator test on that
day.

We need to remove the banner when the day is over.

The most straightforward way to do this is to republish the site at the
start of every day. The gov.uk/alerts code[1] will work out if there are
or aren’t any planned tests to show that day.

1. https://github.com/alphagov/notifications-govuk-alerts/blob/5a274af6d0c202b7bd5e664967b1e113015b9a31/app/models/alerts.py#L38-L44