Remove volume restriction on timeout-notifications
This (mostly) reverts [1], which we will replace with something a
bit better in the next commits. We keep the optimisation [2] and
the change that locks the notifications to update query [3].

We can't really add any tests for the "lock" feature as it's there
to defend against a race condition that's hard to reproduce in a
test. It's not a critical feature, though.

[1]: https://github.com/alphagov/notifications-api/pull/3360/files
[2]: https://github.com/alphagov/notifications-api/pull/3360/files#diff-3f7bdffd1a82923678f3855d8b0d7ad03fe4537224d09ca354999484de9e5420R477
[3]: https://github.com/alphagov/notifications-api/pull/3360/files#diff-3f7bdffd1a82923678f3855d8b0d7ad03fe4537224d09ca354999484de9e5420R484