DRY-up conditionally creating callback tasks
This removes 3 duplicate instances of the same code, which is still
tested implicitly via test_process_ses_receipt_tasks [1]. In the
next commit we'll make this test more explicit, to reflect that it's
now being reused elsewhere and shouldn't change arbitrarily.

We do lose the "print" statement from the command instance of the
code, but I think that's a very tolerable loss.

[1]: https://github.com/alphagov/notifications-api/blob/16ec8ccb8a6e4284eef7ec27c389419b761f39f0/tests/app/celery/test_process_ses_receipts_tasks.py#L94