Add index around key_type on notifications table
Based on [1] - we will need to manually create in Production. The
migration is for other environments / local development.

To optimise the following query, which happens as part of the daily
task to aggregate notification statuses [2]:

    EXPLAIN SELECT
      template_id,
      service_id,
      coalesce(job_id, '00000000-0000-0000-0000-000000000000') AS job_id,
      key_type,
      notification_status,
      count(*) AS notification_count
    FROM notification_history
    WHERE
    created_at >= '2021-12-20' AND
    created_at < '2021-12-21' AND
    notification_type = 'email' AND
    service_id = 'c5956607-20b1-48b4-8983-85d11404e61f' AND
    key_type != 'test'
    GROUP BY 1,2,3,4,5;

The proposed index matches one we have on the notification_history
table [3]. Currently the planner has to use a less comprehensive
index - doesn't include key_type - and the performance is poor.

     GroupAggregate  (cost=383110.51..385168.30 rows=74564 width=72)
       Group Key: template_id, service_id, (COALESCE(job_id, '00000000-0000-0000-0000-000000000000'::uuid)), key_type, notification_status
       ->  Sort  (cost=383110.51..383297.96 rows=74980 width=64)
             Sort Key: template_id, (COALESCE(job_id, '00000000-0000-0000-0000-000000000000'::uuid)), key_type, notification_status
             ->  Index Scan using ix_notifications_service_id_composite on notifications  (cost=0.69..374218.30 rows=74980 width=64)
                   Index Cond: ((service_id = '31814594-61a6-42d6-9041-062e32d439d9'::uuid) AND (notification_type = 'sms'::notification_type) AND (created_at >= '2022-01-04 00:00:00'::timestamp without time zone) AND (created_at < '2022-01-05 00:00:00'::timestamp without time zone))
                   Filter: ((key_type)::text <> 'test'::text)

(The actual query plan is kind of gross [4].)

Note the "Filter" clause at the bottom, which is where the the high
cost is coming from: it's equivalent to a sequential scan [5]. The
difference is clear if we do the same query on the history table,
which is able to make use of the index on all 4 columns [6].

Is there a downside?
--------------------

Every index has a small cost on insertions, etc. that will depend on
the size of the index. It's also a tradeoff: are the benefits good
enough to outweight this small cost? Arguably the answer is "no" as
we're talking about a batch / non-critical task.

However, it will also benefit:

- Deleting notifications [7]. This uses the same set of filters. It's
still a batch task, though, so not a strong example.

- Getting stats for the tiles on the dashboard page [8]. Even though
the "type" filter is in a group by clause, Postgres will still make
use of the index and separately group by the other terms [9].

I think these use cases justify the new index.

What about other indices?
-------------------------

At the moment the query makes use of "service_id_composite", so it's
worth checking if this change will make that index redundant. We can
monitor this by repeatedly running the following:

    select * from pg_stat_user_indexes
    where indexrelname='ix_notifications_service_id_composite';

If the values keep going up after the new index is in place, we know
it's still in use. We won't know for sure until then, so we'll need to
incur the cost of having both indices for a while.

Note: while it's likely the column order does affect the index [10],
here we're just copying the existing one for the history table, since
we know this works well in practice.

[1]: https://github.com/alphagov/notifications-api/pull/2257
[2]: https://github.com/alphagov/notifications-api/blob/b4a11da1ce2121ba05cb5161b9f282565bbf3305/app/dao/fact_notification_status_dao.py#L72-L76
[3]: https://github.com/alphagov/notifications-api/blob/20ead82463a9df1ce3a6325d103a00306609f5fb/app/models.py#L1760-L1764
[4]: https://www.pivotaltracker.com/story/show/180693991/comments/228848846
[5]: https://www.postgresql.org/docs/9.2/using-explain.html
[6]: https://github.com/alphagov/notifications-api/pull/3417
[7]: https://github.com/alphagov/notifications-api/blob/20ead82463a9df1ce3a6325d103a00306609f5fb/app/dao/notifications_dao.py#L403-L406
[8]: https://docs.google.com/document/d/1QZlYKTzFzeUkdDNm8Q_yiHme1edCpPPcA9ak3hvfnUk/edit#heading=h.2rpkvd28brg5
[9]: https://www.pivotaltracker.com/story/show/180693991/comments/228869513
[10]: https://stackoverflow.com/questions/55759506/correct-column-order-for-index-creation-in-postgres