Investigate duplicate tasks
This proves that a task can be duplicated when its duration exceeds
the visibility timeout, **even** if prefetch multiplier is 1.

This is because (for some reason) the worker-level prefetch applies
individually to each worker process [1]:

    "How many messages to prefetch at a time multiplied by the number
     of concurrent processes. The default is 4 (***four messages for
     each process***)."

We could avoid this by having a concurrency of "1" and more instances,
but that would make deployments longer.

Suggested actions
-----------------

Keep the worker as-is: the settings are appropriate for long-running
tasks and do help minimise any duplication.

Continue other work to parallelise and optimise long-running tasks, so
that the visibility timeout is never an issue.

Demo of task duplication
------------------------

In order to do the demo we need:

- 3 tasks and 2 proceses: the hypothesis is that process #1 will pick
up tasks #1 and #2, with task #2 "timing out".

- To poll more often than the timeout, so that process #2 will pick
up (duplicate) task #2 after it finishes #3.

Command to schedule the tasks:

    from app.celery.reporting_tasks import long_task
    [long_task.apply_async(queue='reporting-tasks') for i in [1,2,3]]

Logs that show the duplication:

[2022-01-12 16:50:09,333: INFO/MainProcess] Task long[4c23eaf0-69fc-4951-b215-437d79e30462] received
[2022-01-12 16:50:09,901: INFO/ForkPoolWorker-1] Running long_task
[2022-01-12 16:50:10,208: INFO/MainProcess] Task long[7c22c30d-e61d-4d2d-834c-7dae56b20064] received
[2022-01-12 16:50:10,253: INFO/ForkPoolWorker-2] Running long_task
[2022-01-12 16:50:11,373: INFO/MainProcess] Task long[67e5526d-4a82-430b-8eb5-b8022a20bf70] received
[2022-01-12 16:50:21,410: INFO/MainProcess] Task long[67e5526d-4a82-430b-8eb5-b8022a20bf70] received
[2022-01-12 16:50:30,499: INFO/ForkPoolWorker-3] Running long_task
[2022-01-12 16:50:36,555: INFO/ForkPoolWorker-4] Running long_task

[1]: https://docs.celeryproject.org/en/stable/userguide/configuration.html#std-setting-worker_prefetch_multiplier