Add index to improve page load times for service dashboard
We are seeing slow page load times for the service dashboards, ie
https://www.notifications.service.gov.uk/services/<service_id>

This appears to be caused by slow API request times for the
`get_template_statistics_for_service_by_day` route.

The route appears to be slow due to the query in
`fetch_notification_status_for_service_for_today_and_7_previous_days`
being slow.

That query unions the results from the `ft_notification_status` table
and the `notifications` table. It appears that the query on the
`notifications` table is the slow part:

```
SELECT
	CAST(notifications.notification_type AS TEXT) AS notifications_notification_type,
	notifications.notification_status AS notifications_notification_status,
	notifications.template_id AS notifications_template_id,
	count(*) AS count
FROM notifications
	WHERE
	notifications.created_at >= %(created_at_1)s
	AND notifications.service_id = %(service_id_2)s
	AND notifications.key_type != %(key_type_2)s
GROUP BY notifications.notification_type, notifications.template_id, notifications.notification_status
```

Doing an EXPLAIN ANALYZE on this will vary per service (due to
differing query plans due to different number of notifications per
service) but generally it will do an index scan on
`ix_notifications_service_created_at`. This would do it in two
parts. First it will search the index for all related rows and then
it will go and retrieve those rows. This might be millions of rows
that it needs to retreieve and if these rows are not in memory then
it may need to get these from disk which is particularly slow. We
see in production that often the first page load from 'cold' will
be over 30 seconds but then subsequent page load times shortly
after will be maybe a second or two.

We think one way to solve this is to introduce a new index such
that the index includes all the data in the index itself that this
query needs. An index like this would mean that when searching the
index for relevant rows, there would be no need for the subsequent
step to go and read all the rows (potentially some of them from disk).
This is known as an 'index only scan'
https://www.postgresql.org/docs/12/indexes-index-only-scans.html

The order of the keys in the index is important. We have included
them in the rough order of most differing values to fewest, ie
there are many different service_ids but very few key_types. This
will increase the likelyhood of our queries deciding to use the
new index.

Introducing this index manually and temporarily on staging and doing
some EXPLAIN ANALYZES shows that the slow query
- does an index only scan on our new index
- is dramatically faster (in the region of 90% faster)

Whilst this can't be guaranteed in production, because the data
in staging does not match production, we expect to see somewhat
similar results.

The downside to be aware of is that by adding a new index, everytime
we insert or update a row in the `notifications` table, it will have
to update this index, which takes time. However running a 5 minute
load test on staging at 100 notifications per second, we saw no
difference in response times to create notifications with and
without the new index.

To make this change in production, we must manually run an
`CREATE INDEX CONCURRENTLY` as:
- `CREATE INDEX` would incur downtime
- it is not possible to run `CREATE INDEX CONCURRENTLY`
  using SQLAlchemy

Finally, there may be other options that could solve this problem
but this solution is the proposed because we have good confidence
that it will work. Alternatives that have been considered

Partioning the notifications table:
May work but we aren't sure on what to partition and the amount
of work to figure that out is high and the amount of work to
transition is high (we can't partition the notifications table
itself, you have to create a new partioned table alongside and
mirror traffic)

pg_prewarm:
https://www.postgresql.org/docs/11/pgprewarm.html
This hypothetically could be helpful but we didn't investigate
as it is likely more complex than adding a new index

For further discussion and thoughts, see:
https://docs.google.com/document/d/1QZlYKTzFzeUkdDNm8Q_yiHme1edCpPPcA9ak3hvfnUk/edit#